Data.Delta.Core

Description

Documentation

class Delta delta where

Type class for delta types.

Associated Types

type Base delta

Base type for which delta represents a delta. This is implemented as a type family, so that we can have multiple delta types for the same base type.

Methods

apply :: delta -> Base delta -> Base delta

Apply a delta to the base type.

Whenever the type delta is a Semigroup, we require that

apply (d1 <> d2) = apply d1 . apply d2

This means that deltas are applied right-to-left: d1 is applied after d2.

Whenever the type delta is a Monoid, we require that

apply mempty = id

data NoChange a

The least general delta, where nothing is changed.

Constructors

NoChange

newtype Replace a

Trivial delta type for the type a that replaces the value wholesale.

Constructors

Replace a