Data.Delta.Embedding.Internal

Description

Internal representation of Embedding in terms of state machines.

Documentation

data Machine da db

A state machine that maps deltas to deltas. This machine always carries a state of type Base db around.

Constructors

Machine

Fields

state_ :: !(Base db)

step_ :: (Base da, da) -> (db, Machine da db)

idle :: Delta da => Base da -> Machine da da

Identity machine starting from a base type.

pairMachine :: Machine da1 db1 -> Machine da2 db2 -> Machine (da1, da2) (db1, db2)

Pair two Machines.

fromState :: Delta db => ((Base da, da) -> (Base db, s) -> (db, s)) -> (Base db, s) -> Machine da db

Create a Machine from a specific state s, and the built-in state Base db.