Data.Delta.Set

Description

Delta types for Set.

Single element

data DeltaSet1 a

Delta type for Set where a single element is deleted or added.

Constructors

Insert a

Delete a

The following cancellation laws hold:

apply [Insert a, Delete a] = apply (Insert a)

apply [Insert a, Insert a] = apply (Insert a)

apply [Delete a, Insert a] = apply (Delete a)

apply [Delete a, Delete a] = apply (Delete a)

Multiple elements

data DeltaSet a

Delta type for a Set where collections of elements are inserted or deleted.

diffSet :: Ord a => Set a -> Set a -> DeltaSet a

The smallest delta that changes the second argument to the first argument.

new = apply (diffSet new old) old

diffSet (Set.fromList "ac") (Set.fromList "ab") = deltaSetFromList [Insert 'c', Delete 'b']

listFromDeltaSet :: DeltaSet a -> [DeltaSet1 a]

Flatten a DeltaSet to a list of DeltaSet1.

In the result list, the set of a appearing as Insert a is disjoint from the set of a appearing as Delete a.

deltaSetFromList :: Ord a => [DeltaSet1 a] -> DeltaSet a

Collect insertions or deletions of elements into a DeltaSet.

To save space, combinations of Insert and Delete for the same element are simplified when possible. These simplifications always preserve the property

apply (deltaSetFromList ds) = apply ds